feat: view seed more vert 추가 #8
