+++ seedzip/src/features/View/ThumbnailModal.js
